Migrate Android side & example app to Android Embedding V2
