+++ example/android/app/build.gradle
@@ -33,7 +33,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.jordanalcaraz.audiorecorder.audiorecorderexample"
+        applicationId "com.jordanalcaraz.audiorecorder.audio_recorder_example"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
