--- example/android/app/src/main/java/com/jordanalcaraz/audiorecorder/audio_recorder_example/MainActivity.java
@@ -1,13 +1,0 @@
-package com.jordanalcaraz.audiorecorder.audiorecorderexample;
-
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
-
-public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
-}
